refactor for ease of testing
