+++ extension/.babelrc
@@ -1,7 +1,7 @@
 {
   "sourceMaps": true,
   "presets": [
-    ["env"]
+    ["env", "react"]
   ],
   "plugins": [
     ["transform-react-jsx"],
