+++ extension/src/constants.js
@@ -1,0 +1,10 @@
+// Constants
+export const ToggleType = {
+  ADS: "ads",
+  RATER: "rater"
+};
+
+export const RatingType = {
+  POLITICAL: "political",
+  NORMAL: "normal"
+};
